Factor out logic to find enum's representation type.
Also accept enum representations that fit in long integers.

+++ gnatpolyglot/ada2proxy/src/main/java/com/adacore/gnatpolyglot/ada2proxy/AdaVisitor.java
@@ -734,7 +734,7 @@ public class AdaVisitor extends Libadalang.DefaultVisitor<Void> {
                                     return new EnumLiteral(
                                             lit,
                                             AdaAPI.getName(lit.pDefiningName()),
-                                            lit.pEnumRep().intValue());
+                                            lit.pEnumRep().longValue());
                                 })
                         .toList();
         return new EnumType(decl, AdaAPI.getName(decl.pDefiningName()), enumValues);
